+++ src/main/java/org/mmarini/wheelly/envs/AbstractRobotEnv.java
@@ -30,40 +30,44 @@ package org.mmarini.wheelly.envs;
 
 import io.reactivex.rxjava3.core.Completable;
 import org.mmarini.rl.envs.Environment;
+import org.mmarini.rl.envs.IntSignalSpec;
 import org.mmarini.rl.envs.Signal;
-import org.mmarini.wheelly.apis.RobotControllerApi;
-import org.mmarini.wheelly.apis.RobotStatus;
-import org.mmarini.wheelly.apis.WithRobotStatus;
+import org.mmarini.wheelly.apis.*;
 
 import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.ToDoubleFunction;
 import java.util.function.UnaryOperator;
 
+import static java.lang.Math.round;
 import static java.util.Objects.requireNonNull;
+import static org.mmarini.wheelly.apis.RobotApi.MAX_PPS_SPEED;
+import static org.mmarini.wheelly.apis.Utils.linear;
+import static org.mmarini.wheelly.apis.Utils.normalizeDegAngle;
 
 /**
  * Implements general functionalities of RobotEnvironment
  * <p>
- * The inference is composed by the sequence
+ * The environment should implements
  * <ul>
- *     <li><code>latchStatus</code> stores the eventually composed status for inference processing</li>
- *     <li><code>getSignals</code> retrieves the signals of latched status</li>
- *     <li><code>onAct</code> asks for generation of action
- *     <li><code>processActions</code> processes the result actions
- *     <li><code>getReward</code> generates the result of environment/agent interaction from previous state and current
- *         state (may generate null)
- *     <li><code>onResult</code> observer the result
- *     <li><code>splitStatus</code> split the current status to previous status
+ *     <li><code>onStatus</code> to process the status event</li>
+ *     <li><code>latchStatus</code> to generate eventually composed status and store as current status</li>
+ *     <li><code>getSignals</code> to generate the signals from current composed status</li>
+ *     <li><code>processActions</code> to process the generated actions</li>
+ *     <li><code>splitStatus</code> to split current composed status to previous status for next inference process</li>
  * </ul>
  */
-public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotStatus {
-
+public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotStatus, WithStatusCallback {
+    public static final int MIN_DIRECTION_ACTION = -180;
+    public static final int MAX_DIRECTION_ACTION = 180;
+    public static final int MIN_SENSOR_DIR = -90;
+    public static final int MAX_SENSOR_DIR = 90;
     private final RobotControllerApi controller;
     private final ToDoubleFunction<RobotEnvironment> rewardFunc;
+    private Consumer<RobotStatus> onStatusReady;
     private UnaryOperator<Map<String, Signal>> onAct;
     private Consumer<Environment.ExecutionResult> onResult;
-    private Map<String, Signal> actions;
+    private Map<String, Signal> prevActions;
     private Map<String, Signal> signals0;
     private Consumer<RobotStatus> onInference;
 
@@ -78,6 +82,20 @@ public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotSta
         this.rewardFunc = requireNonNull(rewardFunc);
         controller.setOnInference(this::handleInference);
         controller.setOnLatch(this::latchStatus);
+        controller.setOnStatusReady(this::handleStatus);
+    }
+
+    /**
+     * Returns the delta direction in DEG
+     *
+     * @param actions the actions
+     */
+    int deltaDir(Map<String, Signal> actions) {
+        int action = actions.get("direction").getInt(0);
+        int n = ((IntSignalSpec) getActions().get("direction")).getNumValues();
+        return round(linear(action,
+                0, n - 1,
+                MIN_DIRECTION_ACTION, MAX_DIRECTION_ACTION));
     }
 
     /**
@@ -88,6 +106,10 @@ public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotSta
         return controller;
     }
 
+    public Map<String, Signal> getPrevActions() {
+        return prevActions;
+    }
+
     /**
      * Returns the signals from current status
      */
@@ -95,10 +117,19 @@ public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotSta
 
     /**
      * Processes the inference to produce the behaviour
+     * The inference is composed by the sequence
+     * <ul>
+     *     <li><code>getSignals</code> retrieves the signals of latched status</li>
+     *     <li><code>onAct</code> asks the agent for generation of actions for current state</li>
+     *     <li><code>processActions</code> processes the result actions</li>
+     *     <li><code>getReward</code> generates the rewards from previous state, previous actions and current state</li>
+     *     <li><code>onResult</code> invokes the agent observer for agent training</li>
+     *     <li><code>splitStatus</code> split the current status to previous status</li>
+     * </ul>
      *
      * @param status the current status
      */
-    private void handleInference(RobotStatus status) {
+    protected void handleInference(RobotStatus status) {
         if (onInference != null) {
             onInference.accept(status);
         }
@@ -108,7 +139,7 @@ public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotSta
         if (signals0 != null) {
             double reward = rewardFunc.applyAsDouble(this);
             Environment.ExecutionResult result = new Environment.ExecutionResult(
-                    signals0, actions, reward, signals1, false
+                    signals0, prevActions, reward, signals1, false
             );
             if (onResult != null) {
                 onResult.accept(result);
@@ -116,7 +147,27 @@ public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotSta
         }
         // Split status
         signals0 = signals1;
-        actions = actions1;
+        prevActions = actions1;
+        splitStatus();
+    }
+
+    /**
+     * Handles status event.
+     * Invokes the <code>onStatus</code> method and then the eventually <code>onStatusReady</code> registered call back
+     *
+     * @param status the roboto status
+     */
+    protected void handleStatus(RobotStatus status) {
+        onStatus(status);
+        if (onStatusReady != null) {
+            onStatusReady.accept(status);
+        }
+    }
+
+    public boolean isHalt(Map<String, Signal> actions) {
+        int speedAction = actions.get("speed").getInt(0);
+        int n = ((IntSignalSpec) getActions().get("speed")).getNumValues();
+        return speedAction == n - 1;
     }
 
     /**
@@ -124,18 +175,54 @@ public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotSta
      *
      * @param status the current status
      */
-    protected abstract void latchStatus(RobotStatus status);
+    protected void latchStatus(RobotStatus status) {
+    }
+
+    public int moveDirection(Map<String, Signal> actions, int currentDirection) {
+        int dDir = deltaDir(actions);
+        return normalizeDegAngle(currentDirection + dDir);
+    }
 
     /**
-     * Processes actions
+     * Process status event
+     *
+     * @param status the robot status
+     */
+    protected void onStatus(RobotStatus status) {
+    }
+
+    /**
+     * Processes actions generated by agent
      *
      * @param actions the actions
      */
-    protected abstract void processActions(Map<String, Signal> actions);
+    protected void processActions(Map<String, Signal> actions) {
+        RobotControllerApi controller = getController();
+        int sensorDirection = sensorDir(actions);
+        RobotCommands command = isHalt(actions)
+                ? RobotCommands.haltAndScan(sensorDirection)
+                : RobotCommands.moveAndScan(moveDirection(actions, getRobotStatus().getDirection()),
+                speed(actions),
+                sensorDirection);
+        controller.execute(command);
+    }
 
     @Override
     public Completable readShutdown() {
         return controller.readShutdown();
+    }
+
+    /**
+     * Returns the sensor direction in DEG from actions
+     *
+     * @param actions the actions
+     */
+    public int sensorDir(Map<String, Signal> actions) {
+        int action = actions.get("sensorAction").getInt(0);
+        int n = ((IntSignalSpec) getActions().get("sensorAction")).getNumValues();
+        return round(linear(action,
+                0, n - 1,
+                MIN_SENSOR_DIR, MAX_SENSOR_DIR));
     }
 
     @Override
@@ -154,8 +241,27 @@ public abstract class AbstractRobotEnv implements RobotEnvironment, WithRobotSta
     }
 
     @Override
+    public void setOnStatusReady(Consumer<RobotStatus> callback) {
+        this.onStatusReady = callback;
+    }
+
+    @Override
     public void shutdown() {
         controller.shutdown();
+    }
+
+    public int speed(Map<String, Signal> actions) {
+        int speedAction = actions.get("speed").getInt(0);
+        int n = ((IntSignalSpec) getActions().get("speed")).getNumValues();
+        return round(linear(speedAction,
+                0, n - 1,
+                -MAX_PPS_SPEED, MAX_PPS_SPEED));
+    }
+
+    /**
+     * Splits current composed status to previous status for next inference process
+     */
+    protected void splitStatus() {
     }
 
     @Override