+++ src/main/java/org/mmarini/wheelly/envs/PolarRobotEnv.java
@@ -39,10 +39,8 @@ import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.ToDoubleFunction;
 
-import static java.lang.Math.round;
 import static java.util.Objects.requireNonNull;
-import static org.mmarini.wheelly.apis.RobotApi.MAX_PPS_SPEED;
-import static org.mmarini.wheelly.apis.Utils.*;
+import static org.mmarini.wheelly.apis.Utils.clip;
 import static org.mmarini.yaml.schema.Validator.*;
 
 /**
@@ -83,13 +81,6 @@ public class PolarRobotEnv extends AbstractRobotEnv implements WithPolarMap, Wit
     public static final double MIN_DISTANCE = 0;
     public static final double MAX_DISTANCE = 10;
     public static final int NUM_CONTACT_VALUES = 16;
-
-    public static final int MIN_DIRECTION_ACTION = -180;
-    public static final int MAX_DIRECTION_ACTION = 180;
-    public static final double MIN_SPEED = -1;
-    public static final double MAX_SPEED = 1;
-    public static final int MIN_SENSOR_DIR = -90;
-    public static final int MAX_SENSOR_DIR = 90;
 
     private static final Validator ROBOT_ENV_SPEC = objectPropertiesRequired(Map.of(
                     "objective", object(),
@@ -164,7 +155,7 @@ public class PolarRobotEnv extends AbstractRobotEnv implements WithPolarMap, Wit
     private final double minRadarDistance;
     private CompositeStatus status;
     private CompositeStatus currentStatus;
-    private Consumer<RobotStatus> onStatusReady;
+    private CompositeStatus previousStatus;
 
     /**
      * Creates the controller environment
@@ -199,22 +190,13 @@ public class PolarRobotEnv extends AbstractRobotEnv implements WithPolarMap, Wit
         controller.setOnStatusReady(this::handleStatus);
     }
 
-    /**
-     * Returns the delta direction in DEG
-     *
-     * @param actions the actions
-     */
-    int deltaDir(Map<String, Signal> actions) {
-        int action = actions.get("direction").getInt(0);
-        int n = ((IntSignalSpec) getActions().get("direction")).getNumValues();
-        return round(linear(action,
-                0, n - 1,
-                MIN_DIRECTION_ACTION, MAX_DIRECTION_ACTION));
-    }
-
     @Override
     public Map<String, SignalSpec> getActions() {
         return this.actions;
+    }
+
+    public CompositeStatus getCurrentStatus() {
+        return currentStatus;
     }
 
     public double getMaxRadarDistance() {
@@ -224,6 +206,10 @@ public class PolarRobotEnv extends AbstractRobotEnv implements WithPolarMap, Wit
     @Override
     public PolarMap getPolarMap() {
         return currentStatus.polarMap;
+    }
+
+    public CompositeStatus getPreviousStatus() {
+        return previousStatus;
     }
 
     @Override
@@ -274,14 +260,6 @@ public class PolarRobotEnv extends AbstractRobotEnv implements WithPolarMap, Wit
         return states;
     }
 
-    private void handleStatus(RobotStatus status) {
-        RadarMap newRadarMap = this.status.radarMap.update(status);
-        this.status = this.status.setStatus(status).setRadarMap(newRadarMap);
-        if (onStatusReady != null) {
-            onStatusReady.accept(status);
-        }
-    }
-
     @Override
     protected void latchStatus(RobotStatus ignored) {
         currentStatus = status;
@@ -293,34 +271,9 @@ public class PolarRobotEnv extends AbstractRobotEnv implements WithPolarMap, Wit
     }
 
     @Override
-    protected void processActions(Map<String, Signal> actions) {
-        RobotControllerApi controller = getController();
-        int speedAction = actions.get("speed").getInt(0);
-        int n = ((IntSignalSpec) getActions().get("speed")).getNumValues();
-        if (speedAction == n - 1) {
-            controller.haltRobot();
-        } else {
-            int speed = (int) round(linear(speedAction,
-                    0, n - 1,
-                    -MAX_PPS_SPEED, MAX_PPS_SPEED));
-            int dDir = deltaDir(actions);
-            int direction = normalizeDegAngle(currentStatus.status.getDirection() + dDir);
-            controller.moveRobot(direction, speed);
-        }
-        controller.moveSensor(sensorDir(actions));
-    }
-
-    /**
-     * Returns the sensor direction in DEG from actions
-     *
-     * @param actions the actions
-     */
-    int sensorDir(Map<String, Signal> actions) {
-        int action = actions.get("sensorAction").getInt(0);
-        int n = ((IntSignalSpec) getActions().get("sensorAction")).getNumValues();
-        return round(linear(action,
-                0, n - 1,
-                MIN_SENSOR_DIR, MAX_SENSOR_DIR));
+    protected void onStatus(RobotStatus status) {
+        RadarMap newRadarMap = this.status.radarMap.update(status);
+        this.status = this.status.setStatus(status).setRadarMap(newRadarMap);
     }
 
     @Override
@@ -334,16 +287,16 @@ public class PolarRobotEnv extends AbstractRobotEnv implements WithPolarMap, Wit
     }
 
     @Override
-    public void setOnStatusReady(Consumer<RobotStatus> callback) {
-        onStatusReady = callback;
-    }
-
-    @Override
     public void setOnWriteLine(Consumer<String> callback) {
         getController().setOnWriteLine(callback);
     }
 
-    static class CompositeStatus {
+    @Override
+    protected void splitStatus() {
+        previousStatus = currentStatus;
+    }
+
+    public static class CompositeStatus {
         public final PolarMap polarMap;
         public final RadarMap radarMap;
         public final RobotStatus status;